// src/generated/airline-lookup.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
    "/v1/lists/utilities/airlines": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["airlinesLookup"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        AirlinesLookupResponse: {
            /** @description Repeats associated airline information. */
            AirlineInfo?: {
                /**
                 * @description Returns the specified airline code(s).
                 * @example AI
                 */
                AirlineCode?: string;
                /**
                 * @description The proper business name of the specified airline code.
                 * @example Air India Limited
                 */
                AirlineName?: string;
                /**
                 * @description The common business name of the specified airline code
                 * @example Air India
                 */
                AlternativeBusinessName?: string;
            }[];
            Links?: Omit<components["schemas"]["CommonLinks"], "type">;
        };
        /** @description Array of links for different types of relations, e.g. self, linkTemplate, shop */
        CommonLinks: {
            /**
             * @description Describes relationship between href and current request.
             * @example self
             * @enum {string}
             */
            rel?: "self" | "shop" | "forecast" | "airportsInCity" | "airlineInfo" | "destinations" | "view" | "nextResults" | "prevResults" | "linkTemplate" | "shopTemplate" | "tagLookupLinkTemplate" | "paginatedRequestLinkTemplate";
            /**
             * @description Link to related API request.
             * @example https://api.cert.platform.sabre.com/v1/lists/utilities/airlines?airlinecode=AI
             */
            href?: string;
        }[];
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    airlinesLookup: {
        parameters: {
            query?: {
                /**
                 * @description 2-character IATA airline code.
                 *
                 *     Multiple values accepted in csv format.
                 *     Example: A5,C8,E4
                 * @example A9
                 */
                airlinecode?: string;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Success */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["AirlinesLookupResponse"];
                };
            };
            /**
             * @description Bad Request
             *      - Parameter 'airlinecode' must be a valid value.
             */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
        };
    };
}